requirements.txt
django==4.1.2
Pillow==9.3.0
python-decouple==3.6
djangorestframework==3.14.0
